cudarc::driver

Module result

A thin wrapper around sys.

While all the functions here will return Result, they are mostly all still unsafe because order of operations really matters.

This also only exposes the *_async version of functions because mixing the two is confusing and even more unsafe.

This module also groups functions into sub-modules to make naming easier. For example [sys::cuStreamCreate()] turns into stream::create(), where stream is a module.

Modules

ctx

Context management functions (cuCtx*).

device

Device management functions (cuDevice*).

event

external_memory

function

module

Module management functions (cuModule*).

occupancy

primary_ctx

Primary context management functions (cuDevicePrimaryCtx*).

stream

Stream management functions (cuStream*).

Structs

DriverError

Wrapper around sys::CUresult. See nvidia’s CUresult docs

Functions

free_async⚠

Frees memory with stream ordered semantics.

free_sync⚠

Allocates memory

init

Initializes the CUDA driver API. MUST BE CALLED BEFORE ANYTHING ELSE

launch_cooperative_kernel⚠

Launches a cuda functions

launch_kernel⚠

Launches a cuda functions

malloc_async⚠

Allocates memory with stream ordered semantics.

malloc_managed⚠

Allocates managed memory.

malloc_sync⚠

Allocates memory

mem_advise⚠

Advise about the usage of a given memory range.

mem_get_info

Returns (free, total) memory in bytes.

mem_prefetch_async⚠

See cuda docs **Only available in 12.2+.

memcpy_dtod_async⚠

Copies memory from Device to Device with stream ordered semantics.

memcpy_dtod_sync⚠

Copies memory from Device to Device

memcpy_dtoh_async⚠

Copies memory from Device to Host with stream ordered semantics.

memcpy_dtoh_sync⚠

Copies memory from Device to Host with stream ordered semantics.

memcpy_htod_async⚠

Copies memory from Host to Device with stream ordered semantics.

memcpy_htod_sync⚠

Copies memory from Host to Device

memory_free⚠

Frees device memory.

memset_d8_async⚠

Sets device memory with stream ordered semantics.

memset_d8_sync⚠

Sets device memory with stream ordered semantics.