cudarc::driver::result::external_memory

Function import_external_memory_opaque_fd

pub unsafe fn import_external_memory_opaque_fd(
    fd: RawFd,
    size: u64,
) -> Result<CUexternalMemory, DriverError>

Imports an external memory object, in this case an OpaqueFd.

The memory should be destroyed using destroy_external_memory.

See cuda docs

Safety

size must be the size of the size of the memory object in bytes.