cudarc::driver::result::stream

Function synchronize

pub unsafe fn synchronize(stream: CUstream) -> Result<(), DriverError>

Wait until a stream’s tasks are completed.

See cuda docs

Safety

This should only be called with stream created by create and not already destroyed. This follows default stream semantics, see relevant cuda docs.