cudarc::driver

Module safe

Safe abstractions over crate::driver::result provided by CudaSlice, CudaDevice, CudaStream, and more.

Re-exports

pub use crate::driver::result::DriverError;

Structs

CudaDevice

A wrapper around sys::CUdevice, sys::CUcontext, sys::CUstream, and CudaFunction.

CudaFunction

Wrapper around sys::CUfunction. Used by crate::driver::LaunchAsync.

CudaSlice

Contains a reference counted pointer to both device and host memory allocated for type T.

CudaStream

A wrapper around sys::CUstream that safely ensures null stream is synchronized upon the completion of this stream’s work.

CudaView

A immutable sub-view into a CudaSlice created by CudaSlice::try_slice() or CudaSlice::slice().

CudaViewMut

A mutable sub-view into a CudaSlice created by CudaSlice::try_slice_mut() or CudaSlice::slice_mut().

ExternalMemory

An abstraction for imported external memory.

LaunchConfig

Configuration for result::launch_kernel

MappedBuffer

An abstraction for a mapped buffer for some external memory.

Profiler

Calls profiler_start() in Profiler::new(), and profiler_stop() in Drop.

Traits

DevicePtr

Abstraction over CudaSlice/CudaView

DevicePtrMut

Abstraction over CudaSlice/CudaViewMut

DeviceRepr

Something that can be copied to device memory and turned into a parameter for result::launch_kernel.

DeviceSlice

LaunchAsync

Consumes a CudaFunction to execute asychronously on the device with params determined by generic parameter Params.

ValidAsZeroBits

Marker trait to indicate that the type is valid when all of its bits are set to 0.

Functions

profiler_start

Enables profile collection by the active profiling tool for the current context. If profiling is already enabled, then profiler_start() has no effect. More info in Cuda docs For RAII version see Profiler::new.

profiler_stop

Disables profile collection by the active profiling tool for the current context. If profiling is already disabled, then profiler_stop() has no effect.