cudarc::driver::safe

Struct CudaDevice

pub struct CudaDevice { /* private fields */ }

A wrapper around sys::CUdevice, sys::CUcontext, sys::CUstream, and CudaFunction.

let dev = CudaDevice::new(0).unwrap();

Safety

impl Drop to call all the corresponding resource cleanup methods

Doesn’t impl clone, so you can’t have multiple device pointers hanging around.

Any allocations enforce that self is an Arc, meaning no allocation can outlive the CudaDevice

Implementations

impl CudaDevice

pub unsafe fn upgrade_device_ptr<T>( self: &Arc<Self>, cu_device_ptr: CUdeviceptr, len: usize, ) -> CudaSlice<T>

Creates a CudaSlice from a sys::CUdeviceptr. Useful in conjunction with CudaSlice::leak().

Safety

cu_device_ptr must be a valid allocation

cu_device_ptr must space for len * std::mem::size_of<T>() bytes

The memory may not be valid for type T, so some sort of memset operation should be called on the memory.

impl CudaDevice

pub fn null<T>(self: &Arc<Self>) -> Result<CudaSlice<T>, DriverError>

Allocates an empty CudaSlice with 0 length.

pub unsafe fn alloc<T: DeviceRepr>( self: &Arc<Self>, len: usize, ) -> Result<CudaSlice<T>, DriverError>

Allocates device memory and increments the reference counter of CudaDevice.

Safety

This is unsafe because the device memory is unset after this call.

pub fn alloc_zeros<T: ValidAsZeroBits + DeviceRepr>( self: &Arc<Self>, len: usize, ) -> Result<CudaSlice<T>, DriverError>

Allocates device memory with no associated host memory, and memsets the device memory to all 0s.

Safety

T is marked as ValidAsZeroBits, so the device memory is valid to use

Self is Arc<Self>, and this method increments the rc for self

pub fn memset_zeros<T: ValidAsZeroBits + DeviceRepr, Dst: DevicePtrMut<T>>( self: &Arc<Self>, dst: &mut Dst, ) -> Result<(), DriverError>

Sets all memory to 0 asynchronously.

Safety

T is marked as ValidAsZeroBits, so the device memory is valid to use

Self is Arc<Self>, and this method increments the rc for self

pub fn dtod_copy<T: DeviceRepr, Src: DevicePtr<T>, Dst: DevicePtrMut<T>>( self: &Arc<Self>, src: &Src, dst: &mut Dst, ) -> Result<(), DriverError>

Device to device copy (safe version of result::memcpy_dtod_async).

Panics

If the length of the two values are different

Safety

We are guarunteed that src and dst are pointers to the same underlying type T

Since they are both references, they can’t have been freed

Self is Arc<Self>, and this method increments the rc for self

pub fn htod_copy<T: Unpin + DeviceRepr>( self: &Arc<Self>, src: Vec<T>, ) -> Result<CudaSlice<T>, DriverError>

Takes ownership of the host data and copies it to device data asynchronously.

Safety

Since src is owned by this funcion, it is safe to copy data. Any actions executed after this will take place after the data has been successfully copied.

Self is Arc<Self>, and this method increments the rc for self

pub fn htod_copy_into<T: DeviceRepr + Unpin>( self: &Arc<Self>, src: Vec<T>, dst: &mut CudaSlice<T>, ) -> Result<(), DriverError>

Takes ownership of the host data and copies it to device data asynchronously.

Safety

Since src is owned by this funcion, it is safe to copy data. Any actions executed after this will take place after the data has been successfully copied.

Self is Arc<Self>, and this method increments the rc for self

pub fn htod_sync_copy<T: DeviceRepr>( self: &Arc<Self>, src: &[T], ) -> Result<CudaSlice<T>, DriverError>

Allocates new device memory and synchronously copies data from src into the new allocation.

If you want an asynchronous copy, see CudaDevice::htod_copy().

Safety

Since this function doesn’t own src it is executed synchronously.

Self is Arc<Self>, and this method increments the rc for self

pub fn htod_sync_copy_into<T: DeviceRepr, Dst: DevicePtrMut<T>>( self: &Arc<Self>, src: &[T], dst: &mut Dst, ) -> Result<(), DriverError>

Synchronously copies data from src into the new allocation.

If you want an asynchronous copy, see CudaDevice::htod_copy().

Panics

If the lengths of slices are not equal, this method panics.

Safety

Since this function doesn’t own src it is executed synchronously.

Self is Arc<Self>, and this method increments the rc for self

pub fn dtoh_sync_copy<T: DeviceRepr, Src: DevicePtr<T>>( self: &Arc<Self>, src: &Src, ) -> Result<Vec<T>, DriverError>

Synchronously copies device memory into host memory. Unlike CudaDevice::dtoh_sync_copy_into this returns a Vec<T>.

Safety

Since this function doesn’t own dst (after returning) it is executed synchronously.

Self is Arc<Self>, and this method increments the rc for self

pub fn dtoh_sync_copy_into<T: DeviceRepr, Src: DevicePtr<T>>( self: &Arc<Self>, src: &Src, dst: &mut [T], ) -> Result<(), DriverError>

Synchronously copies device memory into host memory

Use CudaDevice::dtoh_sync_copy if you need Vec<T> and can’t provide a correctly sized slice.

Panics

If the lengths of slices are not equal, this method panics.

Safety

Since this function doesn’t own dst it is executed synchronously.

Self is Arc<Self>, and this method increments the rc for self

pub fn sync_reclaim<T: Clone + Default + DeviceRepr + Unpin>( self: &Arc<Self>, src: CudaSlice<T>, ) -> Result<Vec<T>, DriverError>

Synchronously de-allocates src and converts it into it’s host value. You can just drop the slice if you don’t need the host data.

Safety

Self is Arc<Self>, and this method increments the rc for self

pub fn synchronize(self: &Arc<Self>) -> Result<(), DriverError>

Synchronizes the stream.

impl CudaDevice

pub fn new(ordinal: usize) -> Result<Arc<Self>, DriverError>

Creates a new CudaDevice on device index ordinal.

pub fn new_with_stream(ordinal: usize) -> Result<Arc<Self>, DriverError>

Creates a new CudaDevice on device index ordinal on a non-default stream.

pub fn count() -> Result<i32, DriverError>

pub fn ordinal(&self) -> usize

Get the ordinal index of this CudaDevice.

pub fn name(&self) -> Result<String, DriverError>

Get the name of this device.

pub fn cu_device(&self) -> &CUdevice

Get the underlying sys::CUdevice of this CudaDevice.

Safety

While this function is marked as safe, actually using the returned object is unsafe.

You must not free/release the device pointer, as it is still owned by the CudaDevice.

pub fn cu_primary_ctx(&self) -> &CUcontext

Get the underlying sys::CUcontext of this CudaDevice.

Safety

While this function is marked as safe, actually using the returned object is unsafe.

You must not free/release the context pointer, as it is still owned by the CudaDevice.

pub fn cu_stream(&self) -> &CUstream

Get the underlying sys::CUstream that this CudaDevice executes all of its work on.

Safety

While this function is marked as safe, actually using the returned object is unsafe.

You must not free/release the stream pointer, as it is still owned by the CudaDevice.

pub fn attribute(&self, attrib: CUdevice_attribute) -> Result<i32, DriverError>

Get the value of the specified attribute of this CudaDevice.

impl CudaDevice

pub fn fork_default_stream(self: &Arc<Self>) -> Result<CudaStream, DriverError>

Allocates a new stream that can execute kernels concurrently to the default stream.

The synchronization with default stream happens in code order. See CudaStream docstring.

This stream synchronizes in the following way:

On creation it adds a wait for any existing work on the default work stream to complete

On drop it adds a wait for any existign work on Self to complete to the default stream.

pub fn wait_for( self: &Arc<Self>, stream: &CudaStream, ) -> Result<(), DriverError>

Forces CudaStream to drop, causing the default work stream to block on stream’s completion. This is asynchronous with respect to the host.

impl CudaDevice

pub unsafe fn import_external_memory( self: &Arc<Self>, file: File, size: u64, ) -> Result<ExternalMemory, DriverError>

Import external memory from a File.

Safety

size must be the size of the external memory in bytes.

impl CudaDevice

pub fn has_func(self: &Arc<Self>, module_name: &str, func_name: &str) -> bool

Whether a module and function are currently loaded into the device.

pub fn get_func( self: &Arc<Self>, module_name: &str, func_name: &str, ) -> Option<CudaFunction>

Retrieves a CudaFunction that was registered under module_name and func_name.

impl CudaDevice

pub fn load_ptx( self: &Arc<Self>, ptx: Ptx, module_name: &str, func_names: &[&'static str], ) -> Result<(), DriverError>

Dynamically load a set of crate::driver::CudaFunction from a jit compiled ptx.

ptx contains the compilex ptx

module_name is a unique identifier used to access the module later on with CudaDevice::get_func()

func_names is a slice of function names to load into the module during build.

impl CudaDevice

pub fn bind_to_thread(&self) -> Result<(), DriverError>

Binds the device to the calling thread. You must call this before using the device on a separate thread!