cudarc::driver::safe

Trait ValidAsZeroBits

pub unsafe trait ValidAsZeroBits { }

Marker trait to indicate that the type is valid when all of its bits are set to 0.

Safety

Not all types are valid when all bits are set to 0. Be very sure when implementing this trait!