cudarc::nvrtc

Module result

A thin wrapper around sys providing Results with NvrtcError.

Structs

NvrtcError

Wrapper around sys::nvrtcResult. See nvrtcResult docs

Functions

compile_program⚠

Compiles an already created program. Options should be of the form specified in nvrtc’s supported compiler options.

create_program

Creates a program from source code src. This should be source code from a .cu file.

destroy_program⚠

Releases resources associated with prog.

get_program_log⚠

Extract log from a compiled program.

get_ptx⚠

Extract the ptx associated with prog. Call compile_program() before this.